pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "bitacora_printer"
version = "2.0.0"
description = "Script that generates daily and monk mode bitacoras."
authors = ["Angelo Mosquera <anggelomos@outlook.com>"]
maintainers = ["Angelo Mosquera <anggelomos@outlook.com>"]
readme = "README.md"
homepage = "https://github.com/anggelomos/bitacora-printer"
classifiers = [
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pillow = "*"
python-dateutil = "*"
tickthon = "^0.2.12"
python-weather = "*"
auto-mix-prep = "^0.2.0"
qrcode = "^7.4.2"
nothion = "^0.3.0"
openai = "^1.14.1"
pyinstaller = "^6.12.0"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
mypy = "*"
types-requests = "*"

[tool.poetry.urls]
repository = "https://github.com/anggelomos/bitacora-printer"
